Int -> Float

+++ api/app/models/user.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Boolean, Column, ForeignKey, Integer, String, DateTime
+from sqlalchemy import Boolean, Column, ForeignKey, Integer, String, DateTime, Float
 from sqlalchemy.orm import relationship
 from sqlalchemy.sql import func
 
@@ -10,7 +10,7 @@ class User(Base):
 
     id = Column(Integer, primary_key=True, index=True)
     email = Column(String, unique=True, index=True)
-    credit = Column(Integer, default=0)
+    credit = Column(Float, default=0)
     # role_id = Column(Integer, ForeignKey("roles.id"))
 
     # role = relationship("Role", back_populates="users")
